wip. move WiFiRequestItem to separate folder
