+++ Tests_esp8266/src/LogicProgram.cpp
@@ -19,6 +19,7 @@
 #include "main/Display/LabeledLogicItem.cpp"
 #include "main/Display/MapIOIndicator.cpp"
 #include "main/Display/display.cpp"
+#include "main/LogicProgram/Bindings/WiFiBinding.cpp"
 #include "main/LogicProgram/Controller.cpp"
 #include "main/LogicProgram/InputElement.cpp"
 #include "main/LogicProgram/InputOutputElement.cpp"
@@ -26,7 +27,6 @@
 #include "main/LogicProgram/Inputs/CommonInput.cpp"
 #include "main/LogicProgram/Inputs/CommonTimer.cpp"
 #include "main/LogicProgram/Inputs/InputNC.cpp"
-#include "main/LogicProgram/Bindings/WiFiBinding.cpp"
 #include "main/LogicProgram/Ladder.cpp"
 #include "main/LogicProgram/LadderDesigner.cpp"
 #include "main/LogicProgram/LadderInitial.cpp"
@@ -42,4 +42,6 @@
 #include "main/LogicProgram/Outputs/SetOutput.cpp"
 #include "main/LogicProgram/Serializer/Record.cpp"
 #include "main/LogicProgram/Wire.cpp"
+#include "main/WiFi/WiFiRequestItem.cpp"
+#include "main/WiFi/WiFiService.cpp"
 #include "main/buttons.cpp"
